requirements.txt
snowflake-connector-python
snowflake-snowpark-python
apache-beam[gcp]
python-dotenv
pandas